+++ data/types/common/duration_suite_test.go
@@ -1,0 +1,11 @@
+package commontypes_test
+
+import (
+	"testing"
+
+	"github.com/tidepool-org/platform/test"
+)
+
+func TestSuite(t *testing.T) {
+	test.Test(t)
+}